# credit-service/target/generated-sources/openapi/src/main/java/com/nttdata/creditservice/generated/api/CreditsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nttdata.creditservice.generated.api;

import com.nttdata.creditservice.generated.model.BalanceOperationRequest;
import com.nttdata.creditservice.generated.model.CreateCreditRequest;
import com.nttdata.creditservice.generated.model.CreditResponse;
import com.nttdata.creditservice.generated.model.UpdateCreditRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-27T15:52:46.291209300-05:00[America/Lima]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Credits", description = "Credit operations")
public interface CreditsApi {

    default CreditsApiDelegate getDelegate() {
        return new CreditsApiDelegate() {};
    }

    /**
     * POST /api/credits : Create credit
     *
     * @param createCreditRequest  (required)
     * @return Credit created (status code 201)
     */
    @Operation(
        operationId = "createCredit",
        summary = "Create credit",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Credit created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/credits",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditResponse>> createCredit(
        @Parameter(name = "CreateCreditRequest", description = "", required = true) @Valid @RequestBody Mono<CreateCreditRequest> createCreditRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().createCredit(createCreditRequest, exchange);
    }


    /**
     * DELETE /api/credits/{creditId} : Delete credit
     *
     * @param creditId  (required)
     * @return Credit deleted (status code 204)
     */
    @Operation(
        operationId = "deleteCredit",
        summary = "Delete credit",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Credit deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/credits/{creditId}"
    )
    
    default Mono<ResponseEntity<Void>> deleteCredit(
        @Parameter(name = "creditId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("creditId") String creditId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().deleteCredit(creditId, exchange);
    }


    /**
     * GET /api/credits/{creditId} : Get credit by id
     *
     * @param creditId  (required)
     * @return Credit found (status code 200)
     */
    @Operation(
        operationId = "getCreditById",
        summary = "Get credit by id",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Credit found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/credits/{creditId}",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditResponse>> getCreditById(
        @Parameter(name = "creditId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("creditId") String creditId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getCreditById(creditId, exchange);
    }


    /**
     * GET /api/credits : List credits
     *
     * @param customerId  (optional)
     * @return Credit list (status code 200)
     */
    @Operation(
        operationId = "getCredits",
        summary = "List credits",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Credit list", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CreditResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/credits",
        produces = { "application/json" }
    )
    
    default Mono<ResponseEntity<Flux<CreditResponse>>> getCredits(
        @Parameter(name = "customerId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "customerId", required = false) String customerId,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().getCredits(customerId, exchange);
    }


    /**
     * POST /api/credits/{creditId}/operations : Apply balance operation
     *
     * @param creditId  (required)
     * @param balanceOperationRequest  (required)
     * @return Operation applied (status code 200)
     */
    @Operation(
        operationId = "operateBalance",
        summary = "Apply balance operation",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Operation applied", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/credits/{creditId}/operations",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditResponse>> operateBalance(
        @Parameter(name = "creditId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("creditId") String creditId,
        @Parameter(name = "BalanceOperationRequest", description = "", required = true) @Valid @RequestBody Mono<BalanceOperationRequest> balanceOperationRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().operateBalance(creditId, balanceOperationRequest, exchange);
    }


    /**
     * PUT /api/credits/{creditId} : Update credit
     *
     * @param creditId  (required)
     * @param updateCreditRequest  (required)
     * @return Credit updated (status code 200)
     */
    @Operation(
        operationId = "updateCredit",
        summary = "Update credit",
        tags = { "Credits" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Credit updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreditResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/credits/{creditId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<CreditResponse>> updateCredit(
        @Parameter(name = "creditId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("creditId") String creditId,
        @Parameter(name = "UpdateCreditRequest", description = "", required = true) @Valid @RequestBody Mono<UpdateCreditRequest> updateCreditRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        return getDelegate().updateCredit(creditId, updateCreditRequest, exchange);
    }

}